requirements.txt
Flask>=3.0.0
PyMuPDF>=1.23.0
dashscope>=1.20.0
python-dotenv>=1.0.0
